Move credentials screen to own package
