+++ ui/src/main/java/io/snabble/sdk/ui/payment/payone/sepa/credentials/PayoneSepaActivity.kt
@@ -1,9 +1,9 @@
-package io.snabble.sdk.ui.payment.payone.sepa
+package io.snabble.sdk.ui.payment.payone.sepa.credentials
 
 import androidx.fragment.app.Fragment
 import io.snabble.sdk.ui.BaseFragmentActivity
 
 class PayoneSepaActivity : BaseFragmentActivity() {
 
-    override fun onCreateFragment(): Fragment = PayoneSepaFragment()
+    override fun onCreateFragment(): Fragment = PayoneSepaCredentialsFragment()
 }
